QHull Support Added
This is an initial addition of support for the Qhull library, which
provides support for convex hull, Delaunay triangulation, etc.  It is
intended to use this functionality to build a simple MATLAB-style
'griddata' function to support export of mesh as solid geometry using
IGES.

The intent is to use griddata to obtain a (u,v) representation of each
individual bounded surface on each grain from the (x,y,z) vertex points
from the mesh and then fit a NURBS surface to that using a linear
interpolation algorithm.  The NURBS surfaces will then be trimmed to the
proper bounds and dumped out to an IGES file.